api: rename context to entry
