+++ cubeb-api/src/entry.rs
